+++ test/fixtures/dynamic-paths/nuxt.config.ts
@@ -1,7 +1,9 @@
-import { isNuxtPrepare, projectSuffix, withMatrix } from '../../matrix'
+import { builder, isNuxtPrepare, projectSuffix, withMatrix } from '../../matrix'
 
 export default withMatrix({
   ...(isNuxtPrepare ? {} : { buildDir: `.nuxt-${projectSuffix}` }),
+  // `import.meta.glob` and `?url` asset queries are Vite-only.
+  ...(builder === 'vite' ? {} : { ignore: ['**/dynamic-assets.vue'] }),
   css: ['~/assets/global.css'],
   features: {
     inlineStyles: id => !!id && !id.includes('assets.vue'),
